refactor:DTO location modified
