+++ src/main/java/com/sellingPartners/backEnd/controller/PartnerProfileController.java
@@ -5,6 +5,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -18,11 +19,11 @@ import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.sellingPartners.backEnd.dto.PartnerProfileRequest;
+import com.sellingPartners.backEnd.dto.PartnerProfileResponse;
+import com.sellingPartners.backEnd.dto.PartnerProfileUpdateRequest;
 import com.sellingPartners.backEnd.entity.Category;
 import com.sellingPartners.backEnd.entity.PartnerProfileEntity;
-import com.sellingPartners.backEnd.entity.PartnerProfileRequest;
-import com.sellingPartners.backEnd.entity.PartnerProfileResponse;
-import com.sellingPartners.backEnd.entity.PartnerProfileUpdateRequest;
 import com.sellingPartners.backEnd.service.PartnerProfileService;
 
 import lombok.RequiredArgsConstructor;
@@ -35,7 +36,7 @@ public class PartnerProfileController {
     private final PartnerProfileService partnerProfileService;
     
    
-
+    @Secured("ROLE_USER")
     @PostMapping(value = "", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public ResponseEntity<?> createProfile(
             @RequestParam("title") String title,
