+++ src/main/java/com/sellingPartners/backEnd/controller/ProposalController.java
@@ -1,0 +1,12 @@
+package com.sellingPartners.backEnd.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+import lombok.NoArgsConstructor;
+
+@RestController
+@NoArgsConstructor
+public class ProposalController {
+	
+	
+}
