+++ src/main/java/com/sellingPartners/backEnd/dto/PartnerProfileResponse.java
@@ -1,4 +1,6 @@
-package com.sellingPartners.backEnd.entity;
+package com.sellingPartners.backEnd.dto;
+
+import com.sellingPartners.backEnd.entity.Category;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
